Fancier printf formatting, using field-width specifiers.
Split the 80-column page into left and right halves
with %-40s for the left-justified left half,
then %s for the right half.

Use += to build up the string to print.

Group each of the two arguments to printf with double quotes,
so printf doesn't think multiple words are multiple arguments.